vtt-sync: phase 4 extract map pings to ui/map-pings.js
Move the map ping handlers (~340 lines) out of board-interactions.js
into a new ui/map-pings.js factory. Covers the alt-click ping/focus
interaction, incoming ping dedupe, render lifecycle, and the
sanitize/clone helpers used during persistence and snapshot hashing.

The factory takes its dependencies (board element, ping layer,
view state, board api, persistBoardStateSnapshot, markPingsDirty,
etc.) as accessor callbacks so the existing module-level lifecycle
in board-interactions.js is preserved without behavior changes.

board-interactions.js drops from 20,105 to 19,787 lines. Smoke
tests added in __tests__/map-pings.test.mjs (7 tests, all green).
